crabzilla-core now depends on dagger

+++ crabzilla-core/src/main/java/io/github/crabzilla/vertx/handler.kt
@@ -10,7 +10,7 @@ import java.util.*
 data class CommandExecution(val result: RESULT,
                             val commandId: UUID?,
                             val constraints: List<String> = listOf(),
-                            val uowSequence: Long? = 0L,
+                            val uowSequence: Int? = 0,
                             val unitOfWork: UnitOfWork? = null) : Serializable {
 
   enum class RESULT {
@@ -24,10 +24,9 @@ data class CommandExecution(val result: RESULT,
 
 }
 
+
 interface CommandHandlerService {
 
   fun postCommand(handlerEndpoint: String, command: Command,
                   handler: Handler<AsyncResult<CommandExecution>>)
 }
-
-
